Revert "extend PR #2985 (moved canvas init) to also solve issue #2967 (vr2vr traversal in Oculus Browser) (#2991)"
This reverts commit 5214a16b9aa625463d5b6f2a315b5da8f8977ca5.
