+++ src/index.js
@@ -113,11 +113,3 @@ module.exports = window.AFRAME = {
   utils: utils,
   version: pkg.version
 };
-
-// Enter VR on `vrdisplayactivate`, to handle in-VR traversal much earlier in the lifecycle.
-// This event may also fire in other circumstances (e.g. putting on the Rift headset).
-window.addEventListener('vrdisplayactivate', function (evt) {
-  if (window.AFRAME.scenes.length > 0) {
-    window.AFRAME.scenes[0].enterVR();
-  }
-});
